CuTe

Documentation for CuTe.

CuTe.ArrayEngine

CuTe.CuTeArray

CuTe.ViewEngine

CuTe.coord_to_coord

CuTe.make_fragment_like

CuTe.make_ordered_layout

CuTe.@Layout

CuTe.ArrayEngine — Type

ArrayEngine{T, L} <: Engine{T} <: DenseVector{T}

A owning vector of type T with length L. It is stack-allocated and mutable. It should behaves like a StaticStrideArray with from StrideArrays package.

Examples

function test_alloc()
    x = ArrayEngine{Float32}(one, static(10))
    GC.@preserve x begin sum(ViewEngine(x)) end
end

@test @allocated(test_alloc()) == 0

CuTe.CuTeArray — Type

CuTeArray(engine::DenseVector, layout::Layout)
CuTeArray{T}(::UndefInitializer, layout::StaticLayout)
CuTeArray(ptr::Ptr{T}, layout::StaticLayout)

Create a CuTeArray from an engine and a layout. See also ArrayEngine and ViewEngine.

Examples

julia> slayout = @Layout (5, 2);

julia> array_engine = ArrayEngine{Float32}(one, static(10));

julia> CuTeArray(array_engine, slayout)
5×2 CuTeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{StaticInt{5}, StaticInt{2}}, Tuple{StaticInt{1}, StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

 julia> slayout = @Layout (5,3,2)
(static(5), static(3), static(2)):(static(1), static(5), static(15))

julia> CuTeArray{Float32}(undef, slayout) # uninitialized owning array
5×2 CuTeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 -9.73642f-16   8.09f-43
  8.09f-43     -1.64739f13
  3.47644f36    8.09f-43
  4.5914f-41    0.0
 -9.15084f-21   0.0

julia> A = ones(10); # non-owning array

julia> CuTeArray(pointer(A), slayout)
5×2 CuTeArray{Float64, 2, ViewEngine{Float64, Ptr{Float64}}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

CuTe.ViewEngine — Type

ViewEngine{T, P} <: Engine{T} <: DenseVector{T}

A non-owning view of a memory buffer. P is the type of the pointer.

CuTe.coord_to_coord — Method

Transoform a coordinate in one shape to a coordinate in another shape.

CuTe.make_fragment_like — Method

Make a layout of the same shape with the first mode being col-major, and with the rest following the given order.

CuTe.make_ordered_layout — Method

make_ordered_layout(shape, order)
make_ordered_layout(layout)

Construct a compact layout with the given shape and the stride is following the given order.

Examples

julia> CuTe.make_ordered_layout((3,5), (2, 6))
(3, 5):(static(1), 3)

julia> CuTe.make_ordered_layout((3,5), (10, 2))
(3, 5):(5, static(1))

CuTe.@Layout — Macro

Layout(shape, stride=nothing)

Construct a static layout with the given shape and stride.

Arguments

shape: a tuple of integers or a single integer

stride: a tuple of integers, a single integer, GenColMajor or GenRowMajor