Data Movement

Index

Base.copyto!

Base.copyto! — Method

copyto!(dest::StaticNonOwningArray, src::StaticNonOwningArray)

Copy the contents of src to dest. The function automatically carries out potential vectorization. In particular, while transferring data from global memory to shared memory, it automatically initiates asynchronous copying, if your device supports so.