CuTe

Documentation for CuTe.

CuTe.ArrayEngine

CuTe.CuTeArray

CuTe.Tile

CuTe.ViewEngine

CuTe.blocked_product

CuTe.coord_to_coord

CuTe.logical_divide

CuTe.logical_product

CuTe.make_fragment_like

CuTe.make_ordered_layout

CuTe.raked_product

CuTe.tiled_divide

CuTe.zipped_divide

CuTe.@Layout

CuTe.ArrayEngine — Type

ArrayEngine{T, L} <: Engine{T} <: DenseVector{T}

A owning vector of type T with length L. It is stack-allocated and mutable. It should behaves like a StaticStrideArray with from StrideArrays package.

Examples

function test_alloc()
    x = ArrayEngine{Float32}(one, static(10))
    GC.@preserve x begin sum(ViewEngine(x)) end
end

@test @allocated(test_alloc()) == 0

CuTe.CuTeArray — Type

CuTeArray(engine::DenseVector, layout::Layout)
CuTeArray{T}(::UndefInitializer, layout::StaticLayout)
CuTeArray(ptr::Ptr{T}, layout::StaticLayout)

Create a CuTeArray from an engine and a layout. See also ArrayEngine and ViewEngine.

Examples

julia> slayout = @Layout (5, 2);

julia> array_engine = ArrayEngine{Float32}(one, cosize(slayout));

julia> CuTeArray(array_engine, slayout)
5×2 CuTeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{StaticInt{5}, StaticInt{2}}, Tuple{StaticInt{1}, StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

 julia> slayout = @Layout (5,3,2)
(static(5), static(3), static(2)):(static(1), static(5), static(15))

julia> CuTeArray{Float32}(undef, slayout) # uninitialized owning array
5×2 CuTeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 -9.73642f-16   8.09f-43
  8.09f-43     -1.64739f13
  3.47644f36    8.09f-43
  4.5914f-41    0.0
 -9.15084f-21   0.0

julia> A = ones(10);

julia> CuTeArray(pointer(A), slayout) # create a non-owning array
5×2 CuTeArray{Float64, 2, ViewEngine{Float64, Ptr{Float64}}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

julia> function test_alloc()  # when powered by a ArrayEngine, CuTeArray is stack-allocated
    slayout = @Layout (2, 3)          # and mutable
    x = CuTeArray{Float32}(undef, slayout)
    fill!(x, 1.0f0)
    return sum(x)
end
test_alloc (generic function with 2 methods)

julia> @allocated(test_alloc())
0

CuTe.Tile — Type

A tuple of Layouts or Colons.

CuTe.ViewEngine — Type

ViewEngine{T, P} <: Engine{T} <: DenseVector{T}

A non-owning view of a memory buffer. P is the type of the pointer.

CuTe.blocked_product — Method

blocked_product(tile::Layout, matrix_of_tiles::Layout, coalesce_result::Bool=false)

Compute the blocked product of two layouts. Indexing through the first mode of the new layout corresponds to indexing through the cartesian product of the first mode of tile and the first mode of matrix_of_tiles. Indexing through the second mode is similar. If coalesce_result is true, then the result is coalesced.

julia> tile = make_layout((2, 2), (1, 2));

julia> matrix_of_tiles = make_layout((3, 4), (4, 1));

julia> print_layout(blocked_product(tile, matrix_of_tiles))

CuTe.coord_to_coord — Method

Transoform a coordinate in one shape to a coordinate in another shape.

CuTe.logical_divide — Method

logical_divide(layout::Layout, tile::Tile)

Gather the elements of layout along all modes into blocks according to tile.

julia> raked_prod = make_layout(((3, 2), (4, 2)), ((16, 1), (4, 2)));

julia> print_layout(raked_prod)
((3, 2), (4, 2)):((16, 1), (4, 2))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  5 |  9 | 13 |  3 |  7 | 11 | 15 |
    +----+----+----+----+----+----+----+----+
 2  | 17 | 21 | 25 | 29 | 19 | 23 | 27 | 31 |
    +----+----+----+----+----+----+----+----+
 3  | 33 | 37 | 41 | 45 | 35 | 39 | 43 | 47 |
    +----+----+----+----+----+----+----+----+
 4  |  2 |  6 | 10 | 14 |  4 |  8 | 12 | 16 |
    +----+----+----+----+----+----+----+----+
 5  | 18 | 22 | 26 | 30 | 20 | 24 | 28 | 32 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 38 | 42 | 46 | 36 | 40 | 44 | 48 |
    +----+----+----+----+----+----+----+----+

julia> subtile = (Layout(2, 3), Layout(2, 4)); # gather 2 elements with stride 3 along the first mode
       # and 2 elements with stride 4 along the second mode

julia> print_layout(logical_divide(raked_prod, subtile))
((2, 3), (2, 4)):((1, 16), (2, 4))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  3 |  5 |  7 |  9 | 11 | 13 | 15 |
    +----+----+----+----+----+----+----+----+
 2  |  2 |  4 |  6 |  8 | 10 | 12 | 14 | 16 |
    +----+----+----+----+----+----+----+----+
 3  | 17 | 19 | 21 | 23 | 25 | 27 | 29 | 31 |
    +----+----+----+----+----+----+----+----+
 4  | 18 | 20 | 22 | 24 | 26 | 28 | 30 | 32 |
    +----+----+----+----+----+----+----+----+
 5  | 33 | 35 | 37 | 39 | 41 | 43 | 45 | 47 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 36 | 38 | 40 | 42 | 44 | 46 | 48 |
    +----+----+----+----+----+----+----+----+

CuTe.logical_product — Method

logical_product(A::Layout, B::Layout)

Compute the logical product of two layouts. Indexing through the first mode of the new layout corresponds to indexing through A and indexing through the second mode corresponds to indexing through B.

julia> print_layout(tile)
tile = make_layout((2,2), (1,2));

julia> print_layout(matrix_of_tiles)
(2, 2):(1, 2)
      1   2
    +---+---+
 1  | 1 | 3 |
    +---+---+
 2  | 2 | 4 |
    +---+---+

julia> print_layout(logical_product(tile, matrix_of_tiles));
matrix_of_tiles = make_layout((3,4), (4,1));

CuTe.make_fragment_like — Method

make_fragment_like(::Layout)

Make a compact layout of the same shape with the first mode being col-major, and with the rest following the given order.

CuTe.make_ordered_layout — Method

make_ordered_layout(shape, order)
make_ordered_layout(layout)

Construct a compact layout with the given shape and the stride is following the given order.

Examples

julia> CuTe.make_ordered_layout((3, 5), (2, 6))
(3, 5):(static(1), 3)

julia> CuTe.make_ordered_layout((3, 5), (10, 2))
(3, 5):(5, static(1))

CuTe.raked_product — Method

raked_product(tile::Layout, matrix_of_tiles::Layout, coalesce_result::Bool=false)

The tile is shattered or interleaved with the matrix of tiles.

julia> tile = make_layout((2, 2), (1, 2));

julia> matrix_of_tiles = make_layout((3, 4), (4, 1));

julia> print_layout(raked_product(tile, matrix_of_tiles))
((3, 2), (4, 2)):((16, 1), (4, 2))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  5 |  9 | 13 |  3 |  7 | 11 | 15 |
    +----+----+----+----+----+----+----+----+
 2  | 17 | 21 | 25 | 29 | 19 | 23 | 27 | 31 |
    +----+----+----+----+----+----+----+----+
 3  | 33 | 37 | 41 | 45 | 35 | 39 | 43 | 47 |
    +----+----+----+----+----+----+----+----+
 4  |  2 |  6 | 10 | 14 |  4 |  8 | 12 | 16 |
    +----+----+----+----+----+----+----+----+
 5  | 18 | 22 | 26 | 30 | 20 | 24 | 28 | 32 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 38 | 42 | 46 | 36 | 40 | 44 | 48 |
    +----+----+----+----+----+----+----+----+

CuTe.tiled_divide — Method

tiled_divide(layout::Layout, tile::Tile)

Similar to zipped_divide, but upack the second mode into multiple modes.

CuTe.zipped_divide — Method

zipped_divide(layout::Layout, tile::Tile)

Compute the logical division of layout by tile, then flatten the blocks into single mode and the rest into another mode.

julia> raked_prod = make_layout(((3, 2), (4, 2)), ((16, 1), (4, 2)));

julia> print_layout(raked_prod)

julia> subtile = (Layout(2, 3), Layout(2, 4)); # gather 2 elements with stride 3 along the first mode
       # and 2 elements with stride 4 along the second mode
((3, 2), (4, 2)):((16, 1), (4, 2))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  5 |  9 | 13 |  3 |  7 | 11 | 15 |
    +----+----+----+----+----+----+----+----+
 2  | 17 | 21 | 25 | 29 | 19 | 23 | 27 | 31 |
    +----+----+----+----+----+----+----+----+
 3  | 33 | 37 | 41 | 45 | 35 | 39 | 43 | 47 |
    +----+----+----+----+----+----+----+----+
 4  |  2 |  6 | 10 | 14 |  4 |  8 | 12 | 16 |
    +----+----+----+----+----+----+----+----+
 5  | 18 | 22 | 26 | 30 | 20 | 24 | 28 | 32 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 38 | 42 | 46 | 36 | 40 | 44 | 48 |
    +----+----+----+----+----+----+----+----+

julia> print_layout(zipped_divide(raked_prod, subtile))

CuTe.@Layout — Macro

Layout(shape, stride=nothing)

Construct a static layout with the given shape and stride.

Arguments

shape: a tuple of integers or a single integer

stride: a tuple of integers, a single integer, GenColMajor or GenRowMajor