Shambles

Documentation for Shambles.

Shambles.ArrayEngine

Shambles.CuTeArray

Shambles.Registers

Shambles.Tile

Shambles.ViewEngine

Shambles.blocked_product

Shambles.coord_to_coord

Shambles.get_congr_coord

Shambles.load

Shambles.local_partition

Shambles.local_tile

Shambles.logical_divide

Shambles.logical_product

Shambles.make_fragment_like

Shambles.make_ordered_layout

Shambles.mma

Shambles.raked_product

Shambles.tiled_divide

Shambles.zipped_divide

Shambles.@Layout

Shambles.ArrayEngine — Type

ArrayEngine{T, L} <: Engine{T} <: DenseVector{T}

A owning vector of type T with length L. It is stack-allocated and mutable. It should behaves like a StaticStrideArray with from StrideArrays package.

Examples

function test_alloc()
    x = ArrayEngine{Float32}(one, static(10))
    GC.@preserve x begin sum(ViewEngine(x)) end
end

@test @allocated(test_alloc()) == 0

Shambles.CuTeArray — Type

CuTeArray(engine::DenseVector, layout::Layout)
CuTeArray{T}(::UndefInitializer, layout::StaticLayout)
CuTeArray(ptr::Ptr{T}, layout::StaticLayout)

Create a CuTeArray from an engine and a layout. See also ArrayEngine and ViewEngine.

Examples

julia> slayout = @Layout (5, 2);

julia> array_engine = ArrayEngine{Float32}(one, cosize(slayout));

julia> CuTeArray(array_engine, slayout)
5×2 CuTeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{StaticInt{5}, StaticInt{2}}, Tuple{StaticInt{1}, StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

 julia> slayout = @Layout (5,3,2)
(static(5), static(3), static(2)):(static(1), static(5), static(15))

julia> CuTeArray{Float32}(undef, slayout) # uninitialized owning array
5×2 CuTeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 -9.73642f-16   8.09f-43
  8.09f-43     -1.64739f13
  3.47644f36    8.09f-43
  4.5914f-41    0.0
 -9.15084f-21   0.0

julia> A = ones(10);

julia> CuTeArray(pointer(A), slayout) # create a non-owning array
5×2 CuTeArray{Float64, 2, ViewEngine{Float64, Ptr{Float64}}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}} with indices static(1):static(5)×static(1):static(2):
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

julia> function test_alloc()  # when powered by a ArrayEngine, CuTeArray is stack-allocated
    slayout = @Layout (2, 3)          # and mutable
    x = CuTeArray{Float32}(undef, slayout)
    fill!(x, 1.0f0)
    return sum(x)
end
test_alloc (generic function with 2 methods)

julia> @allocated(test_alloc())
0

Shambles.Registers — Type

Registers{T,S}

A struct that wraps the register file type T and number of register files S.

Shambles.Tile — Type

A tuple of Layouts, Colons or integers.

Shambles.ViewEngine — Type

ViewEngine{T, P} <: Engine{T} <: DenseVector{T}

A non-owning view of a memory buffer. P is the type of the pointer.

Shambles.blocked_product — Method

blocked_product(tile::Layout, matrix_of_tiles::Layout, coalesce_result::Bool=false)

Compute the blocked product of two layouts. Indexing through the first mode of the new layout corresponds to indexing through the cartesian product of the first mode of tile and the first mode of matrix_of_tiles. Indexing through the second mode is similar. If coalesce_result is true, then the result is coalesced.

julia> tile = make_layout((2, 2), (1, 2));

julia> matrix_of_tiles = make_layout((3, 4), (4, 1));

julia> print_layout(blocked_product(tile, matrix_of_tiles))
((2, 3), (2, 4)):((1, 16), (2, 4))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  3 |  5 |  7 |  9 | 11 | 13 | 15 |
    +----+----+----+----+----+----+----+----+
 2  |  2 |  4 |  6 |  8 | 10 | 12 | 14 | 16 |
    +----+----+----+----+----+----+----+----+
 3  | 17 | 19 | 21 | 23 | 25 | 27 | 29 | 31 |
    +----+----+----+----+----+----+----+----+
 4  | 18 | 20 | 22 | 24 | 26 | 28 | 30 | 32 |
    +----+----+----+----+----+----+----+----+
 5  | 33 | 35 | 37 | 39 | 41 | 43 | 45 | 47 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 36 | 38 | 40 | 42 | 44 | 46 | 48 |
    +----+----+----+----+----+----+----+----+

Shambles.coord_to_coord — Method

Transoform a coordinate in one shape to a coordinate in another shape.

Shambles.get_congr_coord — Method

get_congr_coord(l::Layout, index::Integer)

Get the flat congruent coordinate from the physical index index.

Shambles.load — Function

load(src_addr::LLVMPtr, ::LdMatrix) where {T}

Load one or multiple matrices from shared memory to registers. The available LdMatrixs are:

# Type => LLVM intrinsic
"LDSM_U32x1_N" => "llvm.nvvm.ldmatrix.sync.aligned.m8n8.x1.b16.p3"
"LDSM_U32x2_N" => "llvm.nvvm.ldmatrix.sync.aligned.m8n8.x2.b16.p3"
"LDSM_U32x4_N" => "llvm.nvvm.ldmatrix.sync.aligned.m8n8.x4.b16.p3"
"LDSM_U16x2_T" => "llvm.nvvm.ldmatrix.sync.aligned.m8n8.x1.trans.b16.p3"
"LDSM_U16x4_T" => "llvm.nvvm.ldmatrix.sync.aligned.m8n8.x2.trans.b16.p3"
"LDSM_U16x8_T" => "llvm.nvvm.ldmatrix.sync.aligned.m8n8.x4.trans.b16.p3"

You can inspect how many registers are used to store the matrix per thread by

julia> LDSM_U32x4_N()
LD_U32x4_N()

julia> ans.DRegisters
Registers{UInt32, 4}

Note

Would not work with LLVM 14

Shambles.local_partition — Method

local_partition(x::CuTeArray, tile::Tile, coord::Tuple)
local_partition(x::CuTeArray, thread_layout::Layout, thread_id::Integer)

Partition a CuTeArray x into tiles that are parallised over.

Examples

Say we have a CuTeArray x of shape (6, 8) and 4 threads of shape (2, 2). We would like to partition x with the 4 threads and get a view of the entries that the first thread will work on. We can do this by calling local_partition(x, (2, 2), 1).

julia> a = CuTeArray(pointer([i for i in 1:48]), @Layout((6,8)))
6×8 CuTeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{Static.StaticInt{6}, Static.StaticInt{8}}, Tuple{Static.StaticInt{1}, Static.StaticInt{6}}}}:
 1   7  13  19  25  31  37  43
 2   8  14  20  26  32  38  44
 3   9  15  21  27  33  39  45
 4  10  16  22  28  34  40  46
 5  11  17  23  29  35  41  47
 6  12  18  24  30  36  42  48

julia> local_partition(a, (static(2), static(2)), (static(1), static(1)))
3×4 CuTeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{Static.StaticInt{3}, Static.StaticInt{4}}, Tuple{Static.StaticInt{2}, Static.StaticInt{12}}}}:
 1  13  25  37
 3  15  27  39
 5  17  29  41

You can also use a thread layout and index with the thread id to get the tile:

julia> local_partition(a, @Layout((2,2), (1, 2)), 2)
3×4 CuTeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{StaticInt{3}, StaticInt{4}}, Tuple{StaticInt{2}, StaticInt{12}}}}:
 2  14  26  38
 4  16  28  40
 6  18  30  42

julia> local_partition(a, @Layout((2,2), (2, 1)), 2)
3×4 CuTeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{StaticInt{3}, StaticInt{4}}, Tuple{StaticInt{2}, StaticInt{12}}}}:
  7  19  31  43
  9  21  33  45
 11  23  35  47

Shambles.local_tile — Method

local_tile(@nospecialize(x::CuTeArray), tile::Tile, coord::Tuple)

Partition a CuTeArray x into tiles. This is similar to local_partition but tiles are not parallelised.

julia> a = CuTeArray(pointer([i for i in 1:48]), @Layout((6,8)))

julia> local_tile(a, (static(2), static(2)), (static(1), static(1)))
2×2 CuTeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{StaticInt{2}, StaticInt{2}}, Tuple{StaticInt{1}, StaticInt{6}}}}:
 1  7
 2  8

Shambles.logical_divide — Method

logical_divide(layout::Layout, tile::Tile)

Gather the elements of layout along all modes into blocks according to tile.

julia> raked_prod = make_layout(((3, 2), (4, 2)), ((16, 1), (4, 2)));

julia> print_layout(raked_prod)
((3, 2), (4, 2)):((16, 1), (4, 2))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  5 |  9 | 13 |  3 |  7 | 11 | 15 |
    +----+----+----+----+----+----+----+----+
 2  | 17 | 21 | 25 | 29 | 19 | 23 | 27 | 31 |
    +----+----+----+----+----+----+----+----+
 3  | 33 | 37 | 41 | 45 | 35 | 39 | 43 | 47 |
    +----+----+----+----+----+----+----+----+
 4  |  2 |  6 | 10 | 14 |  4 |  8 | 12 | 16 |
    +----+----+----+----+----+----+----+----+
 5  | 18 | 22 | 26 | 30 | 20 | 24 | 28 | 32 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 38 | 42 | 46 | 36 | 40 | 44 | 48 |
    +----+----+----+----+----+----+----+----+

julia> subtile = (Layout(2, 3), Layout(2, 4)); # gather 2 elements with stride 3 along the first mode
       # and 2 elements with stride 4 along the second mode

julia> print_layout(logical_divide(raked_prod, subtile))
((2, 3), (2, 4)):((1, 16), (2, 4))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  3 |  5 |  7 |  9 | 11 | 13 | 15 |
    +----+----+----+----+----+----+----+----+
 2  |  2 |  4 |  6 |  8 | 10 | 12 | 14 | 16 |
    +----+----+----+----+----+----+----+----+
 3  | 17 | 19 | 21 | 23 | 25 | 27 | 29 | 31 |
    +----+----+----+----+----+----+----+----+
 4  | 18 | 20 | 22 | 24 | 26 | 28 | 30 | 32 |
    +----+----+----+----+----+----+----+----+
 5  | 33 | 35 | 37 | 39 | 41 | 43 | 45 | 47 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 36 | 38 | 40 | 42 | 44 | 46 | 48 |
    +----+----+----+----+----+----+----+----+

Shambles.logical_product — Method

logical_product(A::Layout, B::Layout)

Compute the logical product of two layouts. Indexing through the first mode of the new layout corresponds to indexing through A and indexing through the second mode corresponds to indexing through B.

julia> tile = make_layout((2,2), (1,2));

julia> print_layout(tile)
(2, 2):(1, 2)
      1   2
    +---+---+
 1  | 1 | 3 |
    +---+---+
 2  | 2 | 4 |
    +---+---+

julia> matrix_of_tiles = make_layout((3,4), (4,1));

julia> print_layout(matrix_of_tiles)
(3, 4):(4, 1)
       1    2    3    4
    +----+----+----+----+
 1  |  1 |  2 |  3 |  4 |
    +----+----+----+----+
 2  |  5 |  6 |  7 |  8 |
    +----+----+----+----+
 3  |  9 | 10 | 11 | 12 |
    +----+----+----+----+

julia> print_layout(logical_product(tile, matrix_of_tiles));
((2, 2), (3, 4)):((1, 2), (16, 4))
       1    2    3    4    5    6    7    8    9   10   11   12
    +----+----+----+----+----+----+----+----+----+----+----+----+
 1  |  1 | 17 | 33 |  5 | 21 | 37 |  9 | 25 | 41 | 13 | 29 | 45 |
    +----+----+----+----+----+----+----+----+----+----+----+----+
 2  |  2 | 18 | 34 |  6 | 22 | 38 | 10 | 26 | 42 | 14 | 30 | 46 |
    +----+----+----+----+----+----+----+----+----+----+----+----+
 3  |  3 | 19 | 35 |  7 | 23 | 39 | 11 | 27 | 43 | 15 | 31 | 47 |
    +----+----+----+----+----+----+----+----+----+----+----+----+
 4  |  4 | 20 | 36 |  8 | 24 | 40 | 12 | 28 | 44 | 16 | 32 | 48 |
    +----+----+----+----+----+----+----+----+----+----+----+----+

Shambles.make_fragment_like — Method

make_fragment_like(::Layout) -> Layout
make_fragment_like(T, ::CuTeArray) -> CuTeArray

Make a compact layout of the same shape with the first mode being col-major, and with the rest following the given order.

Shambles.make_ordered_layout — Method

make_ordered_layout(shape, order)
make_ordered_layout(layout)

Construct a compact layout with the given shape and the stride is following the given order.

Examples

julia> Shambles.make_ordered_layout((3, 5), (2, 6))
(3, 5):(static(1), 3)

julia> Shambles.make_ordered_layout((3, 5), (10, 2))
(3, 5):(5, static(1))

Shambles.mma — Function

mma(A, B, C, ::MMAOP)

Perform matrix multiply-and-accumulate computation, A*B+C. The available MMAOPs are

"MMAOP_8x8x4_F64F64F64F64_TN" => "llvm.nvvm.mma.m8n8k4.row.col.f64"
"MMAOP_8x8x4_F32F16F16F16_TN" => "llvm.nvvm.mma.m8n8k4.row.col.f32.f16"
"MMAOP_8x8x4_F32F16F16F16_NT" => "llvm.nvvm.mma.m8n8k4.col.row.f32.f16"
"MMAOP_8x8x4_F32F16F16F16_TT" => "llvm.nvvm.mma.m8n8k4.col.col.f32.f16"
"MMAOP_8x8x4_F32F16F16F16_NN" => "llvm.nvvm.mma.m8n8k4.row.row.f32.f16"
"MMAOP_8x8x4_F32F16F16F32_TN" => "llvm.nvvm.mma.m8n8k4.row.col.f32.f32"
"MMAOP_8x8x4_F32F16F16F32_NT" => "llvm.nvvm.mma.m8n8k4.col.row.f32.f32"
"MMAOP_8x8x4_F32F16F16F32_TT" => "llvm.nvvm.mma.m8n8k4.col.col.f32.f32"
"MMAOP_8x8x4_F32F16F16F32_NN" => "llvm.nvvm.mma.m8n8k4.row.row.f32.f32"
"MMAOP_8x8x4_F16F16F16F16_TN" => "llvm.nvvm.mma.m8n8k4.row.col.f16.f16"
"MMAOP_8x8x4_F16F16F16F16_NT" => "llvm.nvvm.mma.m8n8k4.col.row.f16.f16"
"MMAOP_8x8x4_F16F16F16F16_TT" => "llvm.nvvm.mma.m8n8k4.col.col.f16.f16"
"MMAOP_8x8x4_F16F16F16F16_NN" => "llvm.nvvm.mma.m8n8k4.row.row.f16.f16"
"MMAOP_16x8x8_F16F16F16F16_TN" => "llvm.nvvm.mma.m16n8k8.row.col.f16.f16"
"MMAOP_16x8x16_F16F16F16F16_TN" => "llvm.nvvm.mma.m16n8k16.row.col.f16.f16"
"MMAOP_16x8x8_F32F16F16F32_TN" => "llvm.nvvm.mma.m16n8k8.row.col.f32.f32"
"MMAOP_16x8x16_F32F16F16F32_TN" => "llvm.nvvm.mma.m16n8k16.row.col.f32.f32"
"MMAOP_16x8x8_F32BF16BF16F32_TN" => "llvm.nvvm.mma.m16n8k8.row.col.bf16"
"MMAOP_16x8x16_F32BF16BF16F32_TN" => "llvm.nvvm.mma.m16n8k16.row.col.bf16"
"MMAOP_16x8x8_F32TF32TF32F32_TN" => "llvm.nvvm.mma.m16n8k8.row.col.tf32"

You can instantiate any of these MMAOPs and inspect the information about the operation

julia> op =  MMAOP_16x8x8_F32TF32TF32F32_TN()
MMAOP_16x8x8_F32TF32TF32F32_TN()

julia> op.ARegisters        # Register type, and number of registers
Shambles.Registers{UInt32, 4}

julia> op.BRegisters
Shambles.Registers{UInt32, 2}

julia> op.CRegisters
Shambles.Registers{Float32, 4}

Note

Do not use mma with wmma.load together. Their data layouts do not agree. The correct execution chain is ldmatrix + mma.

Shambles.raked_product — Method

raked_product(tile::Layout, matrix_of_tiles::Layout, coalesce_result::Bool=false)

The tile is shattered or interleaved with the matrix of tiles.

julia> tile = make_layout((2, 2), (1, 2));

julia> matrix_of_tiles = make_layout((3, 4), (4, 1));

julia> print_layout(raked_product(tile, matrix_of_tiles))
((3, 2), (4, 2)):((16, 1), (4, 2))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  5 |  9 | 13 |  3 |  7 | 11 | 15 |
    +----+----+----+----+----+----+----+----+
 2  | 17 | 21 | 25 | 29 | 19 | 23 | 27 | 31 |
    +----+----+----+----+----+----+----+----+
 3  | 33 | 37 | 41 | 45 | 35 | 39 | 43 | 47 |
    +----+----+----+----+----+----+----+----+
 4  |  2 |  6 | 10 | 14 |  4 |  8 | 12 | 16 |
    +----+----+----+----+----+----+----+----+
 5  | 18 | 22 | 26 | 30 | 20 | 24 | 28 | 32 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 38 | 42 | 46 | 36 | 40 | 44 | 48 |
    +----+----+----+----+----+----+----+----+

Shambles.tiled_divide — Method

tiled_divide(layout::Layout, tile::Tile)

Similar to zipped_divide, but upack the second mode into multiple modes.

Shambles.zipped_divide — Method

zipped_divide(layout::Layout, tile::Tile)

Compute the logical division of layout by tile, then flatten the blocks into the first mode and the rest into the second mode.

julia> raked_prod = make_layout(((3, 2), (4, 2)), ((16, 1), (4, 2)));

julia> print_layout(raked_prod)
((3, 2), (4, 2)):((16, 1), (4, 2))
       1    2    3    4    5    6    7    8
    +----+----+----+----+----+----+----+----+
 1  |  1 |  5 |  9 | 13 |  3 |  7 | 11 | 15 |
    +----+----+----+----+----+----+----+----+
 2  | 17 | 21 | 25 | 29 | 19 | 23 | 27 | 31 |
    +----+----+----+----+----+----+----+----+
 3  | 33 | 37 | 41 | 45 | 35 | 39 | 43 | 47 |
    +----+----+----+----+----+----+----+----+
 4  |  2 |  6 | 10 | 14 |  4 |  8 | 12 | 16 |
    +----+----+----+----+----+----+----+----+
 5  | 18 | 22 | 26 | 30 | 20 | 24 | 28 | 32 |
    +----+----+----+----+----+----+----+----+
 6  | 34 | 38 | 42 | 46 | 36 | 40 | 44 | 48 |
    +----+----+----+----+----+----+----+----+
julia> subtile = (Layout(2, 3), Layout(2, 4)); # gather 2 elements with stride 3 along the first mode and 2 elements with stride 4 along the second mode

julia> print_layout(zipped_divide(raked_prod, subtile))
((2, 2), (3, 4)):((1, 2), (16, 4))
       1    2    3    4    5    6    7    8    9   10   11   12
    +----+----+----+----+----+----+----+----+----+----+----+----+
 1  |  1 | 17 | 33 |  5 | 21 | 37 |  9 | 25 | 41 | 13 | 29 | 45 |
    +----+----+----+----+----+----+----+----+----+----+----+----+
 2  |  2 | 18 | 34 |  6 | 22 | 38 | 10 | 26 | 42 | 14 | 30 | 46 |
    +----+----+----+----+----+----+----+----+----+----+----+----+
 3  |  3 | 19 | 35 |  7 | 23 | 39 | 11 | 27 | 43 | 15 | 31 | 47 |
    +----+----+----+----+----+----+----+----+----+----+----+----+
 4  |  4 | 20 | 36 |  8 | 24 | 40 | 12 | 28 | 44 | 16 | 32 | 48 |
    +----+----+----+----+----+----+----+----+----+----+----+----+

Shambles.@Layout — Macro

Layout(shape, stride=nothing)

Construct a static layout with the given shape and stride.

Arguments

shape: a tuple of integers or a single integer

stride: a tuple of integers, a single integer, GenColMajor or GenRowMajor