MoYeArray

Index

MoYe.ArrayEngine

MoYe.MoYeArray

MoYe.ViewEngine

MoYe.recast

MoYe.zeros!

MoYe.ViewEngine — Type

ViewEngine{T, P}

A wrapper of a pointer. P is the type of the pointer.

MoYe.ArrayEngine — Type

ArrayEngine{T, L} <: DenseVector{T}

A owning and mutable vector of type T with static length L.

Examples

julia> x = ArrayEngine{Float32}(undef, _3)
3-element ArrayEngine{Float32, 3}:
 -9.8271385f-36
  7.57f-43
 -9.8271385f-36

julia> x[1] = 10f0
10.0f0

julia> x
3-element ArrayEngine{Float32, 3}:
 10.0
  7.57f-43
 -9.8271385f-36

MoYe.MoYeArray — Type

MoYeArray(engine::Engine, layout::Layout)
MoYeArray{T}(::UndefInitializer, layout::StaticLayout)
MoYeArray(ptr, layout::Layout)

Create a MoYeArray from an engine and a layout. See also ArrayEngine and ViewEngine.

Examples

julia> slayout = @Layout (5, 2);

julia> array_engine = ArrayEngine{Float32}(undef, cosize(slayout)); # owning array

julia> MoYeArray(array_engine, slayout)
5×2 MoYeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}}:
 -3.24118f12   0.0
  7.57f-43     0.0
  0.0          0.0
  0.0          0.0
  7.89217f-40  0.0

julia>  MoYeArray{Float32}(undef, slayout)
5×2 MoYeArray{Float32, 2, ArrayEngine{Float32, 10}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}}:
  4.0f-45    7.57f-43
  0.0        0.0
 -1.81623f7  0.0
  7.57f-43   0.0
 -1.81623f7  0.0

julia> A = ones(10);

julia> MoYeArray(pointer(A), slayout) # non-owning array
5×2 MoYeArray{Float64, 2, ViewEngine{Float64, Ptr{Float64}}, Layout{2, Tuple{Static.StaticInt{5}, Static.StaticInt{2}}, Tuple{Static.StaticInt{1}, Static.StaticInt{5}}}}:
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0
 1.0  1.0

julia> function test_alloc()          # when powered by a ArrayEngine, MoYeArray is stack-allocated
    slayout = @Layout (2, 3)          # and mutable
    x = MoYeArray{Float32}(undef, slayout)
    fill!(x, 1.0f0)
    return sum(x)
end
test_alloc (generic function with 2 methods)

julia> @allocated(test_alloc())
0

## Slicing 

Currently don't support slicing syntax [:,2], but we could use `view` to get a view of the array.

julia julia> data = [i for i in 1:164];

julia> a = MoYeArray(data, ((3, 2), (2, _5, _2)), ((4,1), (2, 13, 100))) 6×20 MoYeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{Tuple{Static.StaticInt{3}, Int64}, Tuple{Int64, Static.StaticInt{5}, Static.StaticInt{2}}}, Tuple{Tuple{Int64, Int64}, Tuple{Static.StaticInt{2}, Int64, Int64}}}}: 1 3 14 16 27 29 40 42 53 55 101 103 114 116 127 129 140 142 153 155 5 7 18 20 31 33 44 46 57 59 105 107 118 120 131 133 144 146 157 159 9 11 22 24 35 37 48 50 61 63 109 111 122 124 135 137 148 150 161 163 2 4 15 17 28 30 41 43 54 56 102 104 115 117 128 130 141 143 154 156 6 8 19 21 32 34 45 47 58 60 106 108 119 121 132 134 145 147 158 160 10 12 23 25 36 38 49 51 62 64 110 112 123 125 136 138 149 151 162 164

julia> b = view(a, 2, :) 20-element MoYeArray{Int64, 1, ViewEngine{Int64, Ptr{Int64}}, Layout{1, Tuple{Tuple{Int64, Static.StaticInt{5}, Static.StaticInt{2}}}, Tuple{Tuple{Static.StaticInt{2}, Int64, Int64}}}}: 5 7 18 20 31 33 44 46 57 59 105 107 118 120 131 133 144 146 157 159

MoYe.recast — Function

recast(::Type{NewType}, x::MoYeArray{OldType}) -> MoYeArray{NewType}

Recast the element type of a MoYeArray. This is similar to Base.reinterpret, but dose all the computation at compile time, if possible.

Examples

julia> x = MoYeArray{Int32}(undef, @Layout((2,3)))
2×3 MoYeArray{Int32, 2, ArrayEngine{Int32, 6}, Layout{2, Tuple{Static.StaticInt{2}, Static.StaticInt{3}}, Tuple{Static.StaticInt{1}, Static.StaticInt{2}}}}:
 -1948408944           0  2
         514  -268435456  0

julia> x2 = recast(Int16, x)
4×3 MoYeArray{Int16, 2, ViewEngine{Int16, Ptr{Int16}}, Layout{2, Tuple{Static.StaticInt{4}, Static.StaticInt{3}}, Tuple{Static.StaticInt{1}, Static.StaticInt{4}}}}:
 -23664      0  2
 -29731      0  0
    514      0  0
      0  -4096  0

julia> x3 = recast(Int64, x)
1×3 MoYeArray{Int64, 2, ViewEngine{Int64, Ptr{Int64}}, Layout{2, Tuple{Static.StaticInt{1}, Static.StaticInt{3}}, Tuple{Static.StaticInt{1}, Static.StaticInt{1}}}}:
 2209959748496  -1152921504606846976  2

MoYe.zeros! — Function

zeros!(x::MoYeArray)

Fill x with zeros.