Data Movement

Index

MoYe.cp_async_commit

MoYe.cp_async_wait

MoYe.cucopyto!

MoYe.cucopyto! — Function

cucopyto!(dest::MoYeArray, src::MoYeArray)

Copy the contents of src to dest. The function automatically carries out potential vectorization. In particular, while transferring data from global memory to shared memory, it automatically initiates asynchronous copying, if your device supports so.

Note

It should be used with @gc_preserve if dest or src is powered by an ArrayEngine.

MoYe.cp_async_wait — Function

cp_async_wait(i::Int32)
cp_async_wait()

cp.async.wait.group and cp.async.wait.all.

MoYe.cp_async_commit — Function

cp_async_commit()

cp.async.commit.group.